修复了广角位置修正中算式未进行int转double引起576/564=1的bug

+++ app/src/main/java/com/zaizai1/balltracking/MainActivity.java
@@ -1501,7 +1501,7 @@ public class MainActivity extends Activity implements CvCameraViewListener2, Vie
 
                     double correctedPosition;
                     if(checkBoxWideAngleErrorFix.isChecked()){
-                        correctedPosition=576/564*(position-16)+10;
+                        correctedPosition=(double)576/564*(position-16)+10;
                     }
                     else
                     {
